@@ -13,30 +13,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.1",
-    "@angular/common": "^4.3.1",
-    "@angular/compiler": "^4.3.1",
-    "@angular/core": "^4.3.1",
-    "@angular/forms": "^4.3.1",
-    "@angular/http": "^4.3.1",
-    "@angular/platform-browser": "^4.3.1",
-    "@angular/platform-browser-dynamic": "^4.3.1",
-    "@angular/router": "^4.3.1",
+    "@angular/animations": "^5.2.11",
+    "@angular/common": "^5.2.11",
+    "@angular/compiler": "^5.2.11",
+    "@angular/core": "^5.2.11",
+    "@angular/forms": "^5.2.11",
+    "@angular/http": "^5.2.11",
+    "@angular/platform-browser": "^5.2.11",
+    "@angular/platform-browser-dynamic": "^5.2.11",
+    "@angular/platform-server": "^5.2.11",
+    "@angular/router": "^5.2.11",
     "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.26",
-    "@ngx-translate/core": "^7.0.0",
-    "@ngx-translate/http-loader": "^0.1.0",
+    "@ngx-translate/core": "^9.0.0",
+    "@ngx-translate/http-loader": "^2.0.0",
     "angular2-highcharts": "^0.5.5",
     "core-js": "^2.4.1",
     "cors": "^2.8.4",
     "font-awesome": "^4.7.0",
     "moment": "^2.22.2",
     "ng2-charts": "^1.6.0",
-    "rxjs": "^5.1.0",
+    "ng2-dnd": "^5.0.2",
+    "rxjs": "^5.5.12",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
     "@angular/cli": "^1.3.2",
-    "@angular/compiler-cli": "^4.3.1",
+    "@angular/compiler-cli": "^5.2.11",
     "@angular/language-service": "^4.3.1",
     "@types/jasmine": "2.5.45",
     "@types/node": "~6.0.60",
@@ -52,6 +54,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "^2.4.2"
   }
 }
